pyproject.toml
[project]
name = "telegram-writing-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# On Windows, use no capture to avoid I/O errors during cleanup
addopts = "--capture=no"
